move ColoredFormatter to core.logging
